Initial version that works for SVM processing
