+++ drizzlepac/haputils/astrometric_utils.py
@@ -2435,66 +2435,3 @@ def evaluate_overlap_diffs(diff_dict, limit=1.0):
 
     return verified, max_diff
 
-# -----------------------------------------------------------------------------
-#  Line detection functions
-# -----------------------------------------------------------------------------
-from skimage.transform import hough_line, hough_line_peaks
-from skimage.feature import canny
-
-
-def detect_lines(image):
-    """Detect lines in the input image and return list of line parameters """
-    # extract edges from image for faster line detection
-    edges = canny(image)
-    # turn all sources into filled sources, if they are linear
-    edges = ndimage.binary_erosion(ndimage.binary_dilation(edges, iterations=2), iterations=2)
-
-    # Classic straight-line Hough transform
-    # Set a precision of 0.5 degree.
-    tested_angles = np.linspace(-np.pi / 2, np.pi / 2, 360, endpoint=False)
-    h, theta, d = hough_line(edges, theta=tested_angles)
-
-    lines = []
-    for _, angle, dist in zip(*hough_line_peaks(h, theta, d)):
-        (x0, y0) = dist * np.array([np.cos(angle), np.sin(angle)])
-        lines.append([x0, y0, np.rad2deg(angle),  np.tan(angle + np.pi/2), dist])
-
-    return lines
-
-
-def lines_in_image(image, line_sigma=10.0):
-    """Determine if image is dominated by linear features
-
-    Parameters
-    ----------
-    image : ndarray
-        Background-subtracted image to detect linear features in
-
-    line_sigma : float, optional
-        Detection threshold in sigma used to determine whether
-        the number of features with a given angle is more than
-        the noise.
-
-    Returns
-    -------
-    lines_detected : bool
-        Specifies whether or not image is dominated by linear features
-    """
-    # detect any lines in image
-    lines = detect_lines(image)
-
-    # perform statistical analysis on detected linear features
-    # start by generating a histogram of the angles of all the lines
-    angles = [l[2] for l in lines]
-    angle_bins = np.linspace(-180, 180, 360)  # 1 degree bins
-    ahist = np.histogram(angles, bins=angle_bins)
-
-    # look to see whether the peak angle is dramatically larger
-    # than any other peak
-    # we will use 10-sigma as the threshold by default
-    astats = sigma_clipped_stats(ahist[0], maxiters=1)
-    alimit = astats[2] * line_sigma
-
-    lines_detected = angles.max() > alimit
-
-    return lines_detected
